fixup! JAMES-3823 SMTP Require TLS Option
