+++ server/protocols/protocols-smtp/src/main/java/org/apache/james/smtpserver/tls/SmtpRequireTlsParameterHook.java
@@ -21,57 +21,44 @@ package org.apache.james.smtpserver.tls;
 
 import static org.apache.james.protocols.api.ProtocolSession.State.Transaction;
 
-import java.util.Arrays;
+import java.util.Optional;
 
-import jakarta.mail.MessagingException;
-
+import org.apache.james.core.Username;
 import org.apache.james.protocols.api.ProtocolSession;
 import org.apache.james.protocols.smtp.SMTPRetCode;
 import org.apache.james.protocols.smtp.SMTPSession;
 import org.apache.james.protocols.smtp.hook.HookResult;
 import org.apache.james.protocols.smtp.hook.HookReturnCode;
-import org.apache.james.smtpserver.JamesMessageHook;
-import org.apache.mailet.Attribute;
-import org.apache.mailet.AttributeName;
-import org.apache.mailet.AttributeValue;
-import org.apache.mailet.Mail;
+import org.apache.james.protocols.smtp.hook.MailParametersHook;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SmtpTlsMessageHook implements JamesMessageHook {
-    private static final Logger LOGGER = LoggerFactory.getLogger(SmtpTlsMessageHook.class);
-    public static final String REQUIRETLS = "REQUIRETLS";
-    public static final String TLS_REQUIRED = "TLS-Required";
-
+public class SmtpRequireTlsParameterHook implements MailParametersHook {
+    private static final Logger LOGGER = LoggerFactory.getLogger(SmtpRequireTlsParameterHook.class);
+    private static final String REQUIRETLS = "REQUIRETLS";
     public static final ProtocolSession.AttachmentKey<Boolean> REQUIRETLS_KEY =
         ProtocolSession.AttachmentKey.of(REQUIRETLS, Boolean.class);
 
     @Override
-    public HookResult onMessage(SMTPSession session, Mail mail) {
-        try {
-            session.getAttachment(REQUIRETLS_KEY, Transaction).ifPresent(requireTsl ->
-                mail.setAttribute(new Attribute(AttributeName.of(REQUIRETLS), AttributeValue.of(true))));
-            if (isTlsNotRequired(mail) && !isRequireTlsAttributeContains(mail)) {
-                mail.setAttribute(new Attribute(AttributeName.of(REQUIRETLS), AttributeValue.of(false)));
-            }
-        } catch (MessagingException e) {
-            LOGGER.debug("Incorrect syntax when handling TLS-Required header field", e);
+    public HookResult doMailParameter(SMTPSession session, String paramName, String paramValue) {
+        if (session.getAttachment(REQUIRETLS_KEY, Transaction).isPresent()) {
+            LOGGER.debug("The Mail parameter cannot contain more than one REQUIRETLS parameter at the same time");
             return HookResult.builder()
                 .smtpReturnCode(SMTPRetCode.SYNTAX_ERROR_ARGUMENTS)
                 .hookReturnCode(HookReturnCode.deny())
-                .smtpDescription("Incorrect syntax when handling TLS-Required header field")
+                .smtpDescription("The Mail parameter cannot contain more than one REQUIRETLS parameter at the same time")
                 .build();
+        }
+        if (paramName.equals(REQUIRETLS) && session.isStartTLSSupported() && session.isTLSStarted()) {
+            session.setAttachment(REQUIRETLS_KEY, true, Transaction);
+            String userName = Optional.ofNullable(session.getUsername()).map(Username::asString).orElse("unauthorized");
+            LOGGER.info("SMTP sessionID: {}, User: {}, REQUIRETLS=true", session.getSessionID(), userName);
         }
         return HookResult.DECLINED;
     }
 
-    private static boolean isRequireTlsAttributeContains(Mail mail) {
-        return mail.attributeNames().map(AttributeName::asString)
-            .anyMatch(attributeName -> attributeName.equals(REQUIRETLS));
-    }
-
-    private static boolean isTlsNotRequired(Mail mail) throws MessagingException {
-        String[] headers = mail.getMessage().getHeader(TLS_REQUIRED);
-        return headers != null && headers.length == 1 && (Arrays.asList(headers).contains("No"));
+    @Override
+    public String[] getMailParamNames() {
+        return new String[]{REQUIRETLS};
     }
 }
