+++ server/protocols/protocols-smtp/src/test/resources/smtpserver-requireTls.xml
@@ -13,9 +13,9 @@
     <verifyIdentity>true</verifyIdentity>
     <smtpGreeting>Apache JAMES awesome SMTP Server</smtpGreeting>
     <handlerchain>
-        <handler class="org.apache.james.smtpserver.tls.SmtpTlsEhloHook"/>
-        <handler class="org.apache.james.smtpserver.tls.SmtpTlsParameterHook"/>
-        <handler class="org.apache.james.smtpserver.tls.SmtpTlsMessageHook"/>
+        <handler class="org.apache.james.smtpserver.tls.SmtpRequireTlsEhloHook"/>
+        <handler class="org.apache.james.smtpserver.tls.SmtpRequireTlsParameterHook"/>
+        <handler class="org.apache.james.smtpserver.tls.SmtpRequireTlsMessageHook"/>
         <handler class="org.apache.james.smtpserver.CoreCmdHandlerLoader"/>
     </handlerchain>
     <gracefulShutdown>false</gracefulShutdown>
